+++ udata/features/territories/commands.py
@@ -7,7 +7,10 @@ import lzma
 import os
 import shutil
 import tarfile
+
+from collections import Counter
 from datetime import date
+from string import Formatter
 from urllib import urlretrieve
 
 from udata.models import Dataset, GeoZone, SpatialCoverage
@@ -50,70 +53,83 @@ def migrate_zones_ids():
 
     Should only be run once with the new version of geozones w/ geohisto.
     """
-    counter_datasets = 0
-    counter_zones = 0
-    counter_towns = 0
-    counter_counties = 0
-    counter_regions = 0
-    counter_drom = 0
-    counter_dromcom = 0
-    counter_france = 0
+    counter = Counter()
     drom_zone = GeoZone.objects(id='country-subset:fr:drom').first()
     dromcom_zone = GeoZone.objects(id='country-subset:fr:dromcom').first()
-    france_zone = GeoZone.objects(id='country:fr').first()
-    for dataset in Dataset.objects.all():
-        if dataset.spatial and dataset.spatial.zones:
-            counter_datasets += 1
-            new_zones = []
-            for zone in dataset.spatial.zones:
-                if zone.id.startswith('fr/'):
-                    counter_zones += 1
-                    country, kind, zone_id = zone.id.split('/')
-                    zone_id = zone_id.upper()  # Corsica 2a/b case.
-                    if kind == 'town':
-                        counter_towns += 1
-                        new_zones.append(
-                            GeoZone
-                            .objects(code=zone_id, level='fr:commune')
-                            .valid_at(date.today())
-                            .first())
-                    elif kind == 'county':
-                        counter_counties += 1
-                        new_zones.append(
-                            GeoZone
-                            .objects(code=zone_id, level='fr:departement')
-                            .valid_at(date.today())
-                            .first())
-                    elif kind == 'region':
-                        counter_regions += 1
-                        # Only link to pre-2016 regions which kept the same id.
-                        new_zones.append(
-                            GeoZone
-                            .objects(code=zone_id, level='fr:region')
-                            .first())
-                    else:
-                        new_zones.append(zone)
-                elif zone.id.startswith('country-subset/fr'):
-                    counter_zones += 1
-                    subset, country, kind = zone.id.split('/')
-                    if kind == 'dom':
-                        counter_drom += 1
-                        new_zones.append(drom_zone)
-                    elif kind == 'domtom':
-                        counter_dromcom += 1
-                        new_zones.append(dromcom_zone)
-                elif zone.id.startswith('country/fr'):
-                    counter_zones += 1
-                    counter_france += 1
-                    new_zones.append(france_zone)
+    # Iter over datasets with zones
+    for dataset in Dataset.objects(spatial__zones__gt=[]):
+        counter['datasets'] += 1
+        new_zones = []
+        for zone in dataset.spatial.zones:
+            if zone.id.startswith('fr/'):
+                counter['zones'] += 1
+                country, kind, zone_id = zone.id.split('/')
+                zone_id = zone_id.upper()  # Corsica 2a/b case.
+                if kind == 'town':
+                    counter['towns'] += 1
+                    new_zones.append(
+                        GeoZone
+                        .objects(code=zone_id, level='fr:commune')
+                        .valid_at(date.today())
+                        .first())
+                elif kind == 'county':
+                    counter['counties'] += 1
+                    new_zones.append(
+                        GeoZone
+                        .objects(code=zone_id, level='fr:departement')
+                        .valid_at(date.today())
+                        .first())
+                elif kind == 'region':
+                    counter['regions'] += 1
+                    # Only link to pre-2016 regions which kept the same id.
+                    new_zones.append(
+                        GeoZone
+                        .objects(code=zone_id, level='fr:region')
+                        .first())
+                elif kind == 'epci':
+                    counter['epcis'] += 1
+                    new_zones.append(
+                        GeoZone
+                        .objects(code=zone_id, level='fr:epci')
+                        .valid_at(dataset.created_at.date())
+                        .first())
                 else:
                     new_zones.append(zone)
-            dataset.update(
-                spatial=SpatialCoverage(zones=[z.id for z in new_zones if z]))
-    print('{} datasets and {} zones affected.'.format(
-        counter_datasets, counter_zones))
-    print('{} town zones, {} county zones and {} region zones updated.'.format(
-        counter_towns, counter_counties, counter_regions))
-    print('{} France zones, {} DROM zones, {} DROM-COM zones updated.'.format(
-        counter_france, counter_drom, counter_dromcom))
+            elif zone.id.startswith('country-subset/fr'):
+                counter['zones'] += 1
+                subset, country, kind = zone.id.split('/')
+                if kind == 'dom':
+                    counter['drom'] += 1
+                    new_zones.append(drom_zone)
+                elif kind == 'domtom':
+                    counter['dromcom'] += 1
+                    new_zones.append(dromcom_zone)
+            elif zone.id.startswith('country/'):
+                counter['zones'] += 1
+                counter['countries'] += 1
+                new_zones.append(zone.id.replace('/', ':'))
+            elif zone.id.startswith('country-group/'):
+                counter['zones'] += 1
+                counter['countrygroups'] += 1
+                new_zones.append(zone.id.replace('/', ':'))
+            else:
+                new_zones.append(zone)
+        dataset.update(
+            spatial=SpatialCoverage(
+                granularity=dataset.spatial.granularity,
+                zones=[getattr(z, 'id', z) for z in new_zones if z]
+            )
+        )
+    log.info(Formatter().vformat('''Summary
+    Processed {zones} zones in {datasets} datasets:
+    - {countrygroups} country groups (World/UE)
+    - {countries} countries
+    - France:
+        - {regions} regions
+        - {counties} counties
+        - {epcis} EPCIs
+        - {towns} towns
+        - {drom} DROM
+        - {dromcom} DROM-COM
+    ''', (), counter))
     log.info('Done')