+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Current (in progress)
+
+- Nothing yet
+
 ## 1.1.1 (2017-07-31)
 
 - Updated translations