+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixes some spacing issues on dataset and reuses page buttons
   [#209](https://github.com/etalab/udata-gouvfr/pull/209)
+- Updated some translations
 
 ## 1.1.1 (2017-07-31)