+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   [#209](https://github.com/etalab/udata-gouvfr/pull/209)
 - Fix some wrong spatial coverages
   [#213](https://github.com/etalab/udata-gouvfr/pull/213)
+- Fix translations collision on contact [#211](https://github.com/etalab/udata-gouvfr/pull/211) [#212](https://github.com/etalab/udata-gouvfr/pull/212)
 - Updated some translations
 
 ## 1.1.1 (2017-07-31)