+++ CHANGELOG.md
@@ -14,6 +14,8 @@
   [1094](https://github.com/opendatateam/udata/pull/1094)
 - Prevent the "Bad request error" happening on search but only on some servers
   [#1097](https://github.com/opendatateam/udata/pull/1097)
+- Migrate spatial granularities to new identifiers
+  [#1079](https://github.com/opendatateam/udata/pull/1079)
 
 ## 1.1.1 (2017-07-31)