+++ CHANGELOG.md
@@ -16,7 +16,7 @@
   [#1097](https://github.com/opendatateam/udata/pull/1097)
 - Migrate spatial granularities to new identifiers
 - Migrate remaining legacy spatial identifiers
-  [#1079](https://github.com/opendatateam/udata/pull/1079)
+  [#1080](https://github.com/opendatateam/udata/pull/1080)
 
 ## 1.1.1 (2017-07-31)