+++ CHANGELOG.md
@@ -14,6 +14,9 @@
   [1094](https://github.com/opendatateam/udata/pull/1094)
 - Prevent the "Bad request error" happening on search but only on some servers
   [#1097](https://github.com/opendatateam/udata/pull/1097)
+- Migrate spatial granularities to new identifiers
+- Migrate remaining legacy spatial identifiers
+  [#1080](https://github.com/opendatateam/udata/pull/1080)
 
 ## 1.1.1 (2017-07-31)